UI change, 2 FXML added plus controller class change
