+++ pom.xml
@@ -60,7 +60,7 @@
                         <!-- Default configuration for running with: mvn clean javafx:run -->
                         <id>default-cli</id>
                         <configuration>
-                            <mainClass>com.example.todolist/com.example.todolist.ProgramStart</mainClass>
+                            <mainClass>com.example.todolist/com.example.JDesk.ProgramStart</mainClass>
                             <launcher>app</launcher>
                             <jlinkZipName>app</jlinkZipName>
                             <jlinkImageName>app</jlinkImageName>
