+++ src/main/java/com/example/JDesk/UIController.java
@@ -1,0 +1,57 @@
+package com.example.JDesk;
+
+import javafx.event.ActionEvent;
+import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Node;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
+import javafx.scene.control.TextField;
+import javafx.stage.Stage;
+
+import java.io.IOException;
+
+
+public class UIController {
+
+    @FXML
+    private TextField NameField;
+    @FXML
+    private TextField PasswordField;
+    @FXML
+    private TextField EmailField;
+    @FXML
+    private Label NameLabel;
+    @FXML
+    private Label PasswordLabel;
+    @FXML
+    private Label EmailLabel;
+    @FXML
+    private Button AlreadyClick;
+    @FXML
+    private Button RegisterClick;
+    @FXML
+    private Button SignButton;
+    @FXML
+    private Label Logo;
+
+
+
+    @FXML
+public void SignClicked(ActionEvent event) throws Exception
+
+    {
+        Parent root = FXMLLoader.load (getClass().getResource("RegisterUI.fxml"));
+        Scene scene = new Scene(root);
+        Stage stage = (Stage)((Node) event.getSource()).getScene().getWindow();
+        stage.setScene(scene);
+        stage.show();
+
+
+    }
+
+
+
+}
