+++ src/main/resources/com/example/JDesk/LogoUI.fxml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="837.0"
+            prefWidth="1084.0" xmlns="http://javafx.com/javafx/20.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.JDesk.UIController">
+   <VBox layoutX="542.0" layoutY="312.0" prefHeight="118.0" prefWidth="248.0">
+      <Label alignment="CENTER" prefHeight="44.0" prefWidth="241.0" text="Welcome To">
+         <font>
+            <Font size="36.0"/>
+         </font>
+      </Label>
+      <Label alignment="CENTER" contentDisplay="CENTER" prefHeight="72.0" prefWidth="178.0" text="JDesk"
+             textAlignment="CENTER">
+         <font>
+            <Font size="43.0"/>
+         </font>
+      </Label>
+   </VBox>
+   <Button fx:id="SignButton" layoutX="569.0" layoutY="518.0" mnemonicParsing="false" onAction="#SignClicked"
+           opacity="0.62" prefHeight="26.0" prefWidth="217.0" text="Sign-In/Sing-Up"/>
+</AnchorPane>
