--- src/main/resources/com/example/todolist/LogoUI.fxml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-
-
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="590.0" prefWidth="685.0" xmlns="http://javafx.com/javafx/20.0.1" xmlns:fx="http://javafx.com/fxml/1">
-   <children>
-      <Label alignment="CENTER" layoutX="174.0" layoutY="149.0" prefHeight="77.0" prefWidth="267.0" text="SkyDesk" textAlignment="CENTER" AnchorPane.bottomAnchor="174.0" AnchorPane.leftAnchor="174.0" AnchorPane.rightAnchor="159.0" AnchorPane.topAnchor="149.0">
-         <font>
-            <Font size="48.0" />
-         </font>
-      </Label>
-      <Label layoutX="329.0" layoutY="252.0" prefHeight="51.0" prefWidth="91.0" text="Welcome to" AnchorPane.bottomAnchor="397.0" AnchorPane.leftAnchor="629.0" AnchorPane.rightAnchor="180.0" AnchorPane.topAnchor="252.0" />
-   </children>
-</AnchorPane>
