--- src/main/resources/com/example/todolist/SkyUI.fxml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-        <?import javafx.scene.control.Button?>
-        <?import javafx.scene.control.Label?>
-        <?import javafx.scene.control.TextField?>
-        <?import javafx.scene.layout.AnchorPane?>
-        <?import javafx.scene.text.Font?>
-
-
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0"
-            prefWidth="600.0" xmlns="http://javafx.com/javafx/20.0.1" xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="com.example.todolist.UIController">
-    <Label fx:id="Logo" alignment="CENTER" layoutX="197.0" layoutY="136.0" prefHeight="79.0" prefWidth="200.0"
-           text="SkyDesk" AnchorPane.bottomAnchor="171.0" AnchorPane.leftAnchor="200.0" AnchorPane.rightAnchor="200.0"
-           AnchorPane.topAnchor="150.0">
-        <font>
-            <Font name="System Bold Italic" size="36.0"/>
-        </font>
-    </Label>
-    <Button fx:id="SignButton" onAction="#SignClicked" layoutX="139.0" layoutY="236.0" mnemonicParsing="false" prefHeight="26.0"
-            prefWidth="135.0" text="Sign-In/Sign-Up" />
-    <AnchorPane prefHeight="400.0" prefWidth="600.0">
-        <Label fx:id="NameLabel" layoutX="94.0" layoutY="50.0" prefHeight="35.0" prefWidth="109.0" text="Name"/>
-        <Label fx:id="PasswordLabel" layoutX="94.0" layoutY="91.0" prefHeight="26.0" prefWidth="109.0"
-               text="Password"/>
-        <Button fx:id="RegisterClick" layoutX="211.0" layoutY="200.0" mnemonicParsing="false" prefHeight="26.0"
-                prefWidth="135.0" text="Register"/>
-        <Label fx:id="EmailLabel" layoutX="94.0" layoutY="128.0" prefHeight="26.0" prefWidth="65.0" text="Email"/>
-        <Button fx:id="AlreadyClick" layoutX="363.0" layoutY="201.0" mnemonicParsing="false" opacity="0.38"
-                prefHeight="26.0" prefWidth="124.0" text="Already User?"/>
-        <TextField fx:id="NameField" layoutX="202.0" layoutY="55.0"/>
-        <TextField fx:id="EmailField" layoutX="202.0" layoutY="128.0"/>
-        <TextField fx:id="PasswordField" layoutX="202.0" layoutY="91.0"/>
-    </AnchorPane>
-</AnchorPane>
